+++ sdk/log/bench_test.go
@@ -16,59 +16,77 @@ import (
 func BenchmarkProcessor(b *testing.B) {
 	for _, tc := range []struct {
 		name string
-		f    func() Processor
+		f    func() []LoggerProviderOption
 	}{
 		{
 			name: "Simple",
-			f: func() Processor {
-				return NewSimpleProcessor(noopExporter{})
+			f: func() []LoggerProviderOption {
+				return []LoggerProviderOption{WithProcessor(NewSimpleProcessor(noopExporter{}))}
 			},
 		},
 		{
 			name: "Batch",
-			f: func() Processor {
-				return NewBatchProcessor(noopExporter{})
+			f: func() []LoggerProviderOption {
+				return []LoggerProviderOption{WithProcessor(NewBatchProcessor(noopExporter{}))}
 			},
 		},
 		{
 			name: "SetTimestampSimple",
-			f: func() Processor {
-				return timestampDecorator{NewSimpleProcessor(noopExporter{})}
+			f: func() []LoggerProviderOption {
+				return []LoggerProviderOption{
+					WithProcessor(timestampProcessor{}),
+					WithProcessor(NewSimpleProcessor(noopExporter{})),
+				}
 			},
 		},
 		{
 			name: "SetTimestampBatch",
-			f: func() Processor {
-				return timestampDecorator{NewBatchProcessor(noopExporter{})}
+			f: func() []LoggerProviderOption {
+				return []LoggerProviderOption{
+					WithProcessor(timestampProcessor{}),
+					WithProcessor(NewBatchProcessor(noopExporter{})),
+				}
 			},
 		},
 		{
 			name: "AddAttributesSimple",
-			f: func() Processor {
-				return attrAddDecorator{NewSimpleProcessor(noopExporter{})}
+			f: func() []LoggerProviderOption {
+				return []LoggerProviderOption{
+					WithProcessor(attrAddProcessor{}),
+					WithProcessor(NewSimpleProcessor(noopExporter{})),
+				}
 			},
 		},
 		{
 			name: "AddAttributesBatch",
-			f: func() Processor {
-				return attrAddDecorator{NewBatchProcessor(noopExporter{})}
+			f: func() []LoggerProviderOption {
+				return []LoggerProviderOption{
+					WithProcessor(attrAddProcessor{}),
+					WithProcessor(NewBatchProcessor(noopExporter{})),
+				}
 			},
 		},
 		{
 			name: "SetAttributesSimple",
-			f: func() Processor {
-				return attrSetDecorator{NewSimpleProcessor(noopExporter{})}
+			f: func() []LoggerProviderOption {
+				return []LoggerProviderOption{
+					WithProcessor(attrSetDecorator{}),
+					WithProcessor(NewSimpleProcessor(noopExporter{})),
+				}
 			},
 		},
 		{
 			name: "SetAttributesBatch",
-			f: func() Processor {
-				return attrSetDecorator{NewBatchProcessor(noopExporter{})}
+			f: func() []LoggerProviderOption {
+				return []LoggerProviderOption{
+					WithProcessor(attrSetDecorator{}),
+					WithProcessor(NewBatchProcessor(noopExporter{})),
+				}
 			},
 		},
 	} {
 		b.Run(tc.name, func(b *testing.B) {
-			provider := NewLoggerProvider(WithProcessor(tc.f()))
+			provider := NewLoggerProvider(tc.f()...)
 			b.Cleanup(func() { assert.NoError(b, provider.Shutdown(context.Background())) })
 			logger := provider.Logger(b.Name())
 
@@ -91,32 +109,59 @@ func BenchmarkProcessor(b *testing.B) {
 	}
 }
 
-type timestampDecorator struct {
-	Processor
-}
+type timestampProcessor struct{}
 
-func (e timestampDecorator) OnEmit(ctx context.Context, r Record) error {
-	r = r.Clone()
+func (p timestampProcessor) OnEmit(ctx context.Context, r *Record) error {
 	r.SetObservedTimestamp(time.Date(1988, time.November, 17, 0, 0, 0, 0, time.UTC))
-	return e.Processor.OnEmit(ctx, r)
+	return nil
 }
 
-type attrAddDecorator struct {
-	Processor
+func (p timestampProcessor) Enabled(context.Context, Record) bool {
+	return true
 }
 
-func (e attrAddDecorator) OnEmit(ctx context.Context, r Record) error {
-	r = r.Clone()
+func (p timestampProcessor) Shutdown(ctx context.Context) error {
+	return nil
+}
+
+func (p timestampProcessor) ForceFlush(ctx context.Context) error {
+	return nil
+}
+
+type attrAddProcessor struct{}
+
+func (p attrAddProcessor) OnEmit(ctx context.Context, r *Record) error {
 	r.AddAttributes(log.String("add", "me"))
-	return e.Processor.OnEmit(ctx, r)
+	return nil
 }
 
-type attrSetDecorator struct {
-	Processor
+func (p attrAddProcessor) Enabled(context.Context, Record) bool {
+	return true
 }
 
-func (e attrSetDecorator) OnEmit(ctx context.Context, r Record) error {
-	r = r.Clone()
+func (p attrAddProcessor) Shutdown(ctx context.Context) error {
+	return nil
+}
+
+func (p attrAddProcessor) ForceFlush(ctx context.Context) error {
+	return nil
+}
+
+type attrSetDecorator struct{}
+
+func (p attrSetDecorator) OnEmit(ctx context.Context, r *Record) error {
 	r.SetAttributes(log.String("replace", "me"))
-	return e.Processor.OnEmit(ctx, r)
+	return nil
+}
+
+func (p attrSetDecorator) Enabled(context.Context, Record) bool {
+	return true
+}
+
+func (p attrSetDecorator) Shutdown(ctx context.Context) error {
+	return nil
+}
+
+func (p attrSetDecorator) ForceFlush(ctx context.Context) error {
+	return nil
 }